refactor: タブのアイコンをコンポーネント化
